change some stuff
