+++ main.py
@@ -1,41 +1,52 @@
-import pickle
-import numpy as np
+from fastapi import FastAPI, Form, Request
+from fastapi.responses import HTMLResponse
+from fastapi.staticfiles import StaticFiles
+from fastapi.templating import Jinja2Templates
+from prediction import heart_attack
+
+app = FastAPI()
+
+# Mount the static files directory
+app.mount("/static", StaticFiles(directory="static"), name="static")
+
+# Set up Jinja2 templates
+templates = Jinja2Templates(directory="templates")
 
 
-with open('features.pkl', 'rb') as f:
-    features = pickle.load(f)
-
-with open('scaler.pkl', 'rb') as f:
-    scaler = pickle.load(f)
-
-with open('model.pkl', 'rb') as f:
-    model = pickle.load(f)
-
-
-def heart_attack(age, restingBp, cholesterol, maxhr, sex, chest_pain_type, exercise_agna, st_slope, fasting_bs):
-    pred_test = np.zeros(len(features.columns))
-
-    chest_pain_type_index = np.where(features.columns == chest_pain_type)[0][0]
-    sex_index = np.where(features.columns == sex)[0][0]
-    exercise_agna_index = np.where(features.columns == exercise_agna)[0][0]
-    st_slope_index = np.where(features.columns == st_slope)[0][0]
-    fasting_bs_index = np.where(features.columns == fasting_bs)[0][0]
-
-    pred_test[0] = age
-    pred_test[1] = restingBp
-    pred_test[2] = cholesterol
-    pred_test[3] = maxhr
-    pred_test[sex_index] = 1
-    pred_test[chest_pain_type_index] = 1
-    pred_test[exercise_agna_index] = 1
-    pred_test[st_slope_index] = 1
-    pred_test[fasting_bs_index] = 1
-
-   
-    scaled_prediction = scaler.transform([pred_test])
-
-    return model.predict(scaled_prediction)
+class HeartDiseaseData:
+    def __init__(self, age: int, restingBP: int, cholesterol: int, maxhr: int, sex: str, chest_pain_type: str, exercise_agna: str, st_slope: str, fasting_bs: str):
+        self.age = age
+        self.restingBP = restingBP
+        self.cholesterol = cholesterol
+        self.maxhr = maxhr
+        self.sex = sex
+        self.chest_pain_type = chest_pain_type
+        self.exercise_agna = exercise_agna
+        self.st_slope = st_slope
+        self.fasting_bs = fasting_bs
 
 
-prediction = heart_attack(54, 140, 289, 172, 'F', 'ATA', 'N', 'Up', '0')
-print(prediction)
+@app.get("/", response_class=HTMLResponse)
+async def index(request: Request):
+    return templates.TemplateResponse("index.html", {"request": request})
+
+
+@app.post("/predict/", response_class=HTMLResponse)
+async def predict(request: Request, age: int = Form(...), restingBP: int = Form(...), cholesterol: int = Form(...),
+                  maxhr: int = Form(...), sex: str = Form(...), chest_pain_type: str = Form(...),
+                  exercise_agna: str = Form(...), st_slope: str = Form(...), fasting_bs: str = Form(...)):
+
+    # Call the heart_attack function with the received data
+    result = heart_attack(age, restingBP, cholesterol, maxhr,
+                          sex, chest_pain_type, exercise_agna, st_slope, fasting_bs)
+
+    # Process the result or do something with it
+    print(result)
+    print(type(result))
+
+    if result == 0:
+        template_name = "no_heart_disease.html"
+    elif result == 1:
+        template_name = "heart_disease.html"
+
+    return templates.TemplateResponse(template_name, {"request": request})
