+++ requirements.txt
@@ -1,168 +1,20 @@
-absl-py==1.4.0
-altair==5.0.1
-annotated-types==0.6.0
-anyio==4.3.0
-asttokens==2.2.1
-astunparse==1.6.3
-attrs==23.1.0
-backcall==0.2.0
-blinker==1.6.2
-boto3==1.28.19
-botocore==1.31.19
-cachetools==5.3.1
-certifi==2023.7.22
-charset-normalizer==3.2.0
-click==8.1.6
-colorama==0.4.6
-comm==0.1.4
-contourpy==1.2.0
-cycler==0.12.1
-debugpy==1.6.7.post1
-decorator==5.1.1
-dnspython==2.6.1
-email_validator==2.1.1
-enchant==0.0.1
-et-xmlfile==1.1.0
-executing==1.2.0
-fastapi==0.110.1
-filelock==3.12.2
-flatbuffers==23.5.26
-fonttools==4.44.3
-fsspec==2023.6.0
-gast==0.4.0
-gitdb==4.0.10
-GitPython==3.1.32
-google-api-core==2.15.0
-google-api-python-client==2.115.0
-google-auth==2.22.0
-google-auth-httplib2==0.2.0
-google-auth-oauthlib==1.0.0
-google-pasta==0.2.0
-googleapis-common-protos==1.62.0
-grpcio==1.57.0
-h11==0.14.0
-h5py==3.9.0
-httpcore==1.0.5
-httplib2==0.22.0
-httptools==0.6.1
-httpx==0.27.0
-huggingface-hub==0.16.4
-idna==3.4
-importlib-metadata==6.8.0
-ipykernel==6.25.1
-ipython==8.14.0
-itsdangerous==2.1.2
-jedi==0.19.0
 Jinja2==3.1.2
-jmespath==1.0.1
-joblib==1.3.1
-jsonschema==4.19.0
-jsonschema-specifications==2023.7.1
-jupyter_client==8.3.0
-jupyter_core==5.3.1
-keras==2.13.1
-kiwisolver==1.4.5
-libclang==16.0.6
-Markdown==3.4.4
-markdown-it-py==3.0.0
-MarkupSafe==2.1.3
+fastapi==0.110.1
+
+
+
 matplotlib==3.8.2
-matplotlib-inline==0.1.6
-mdurl==0.1.2
-mpmath==1.3.0
-mysql-connector-python==8.2.0
-nest-asyncio==1.5.7
-networkx==3.1
-nltk==3.8.1
+
 numpy==1.24.3
-oauthlib==3.2.2
-openpyxl==3.1.2
-opt-einsum==3.3.0
-orjson==3.10.0
-packaging==23.1
+
 pandas==2.0.3
-parso==0.8.3
-pickleshare==0.7.5
-Pillow==9.5.0
-platformdirs==3.10.0
-prompt-toolkit==3.0.39
-protobuf==4.21.12
-psutil==5.9.5
-pure-eval==0.2.2
-pyarrow==12.0.1
-pyasn1==0.5.0
-pyasn1-modules==0.3.0
+
 pydantic==2.6.4
-pydantic-extra-types==2.6.0
-pydantic-settings==2.2.1
-pydantic_core==2.16.3
-pydeck==0.8.0
-Pygments==2.16.1
-Pympler==1.0.1
-pyparsing==3.1.1
-pyspellchecker==0.7.2
-pyTelegramBotAPI==4.14.0
-python-dateutil==2.8.2
-python-dotenv==1.0.1
-python-multipart==0.0.9
-pytz==2023.3
-pytz-deprecation-shim==0.1.0.post0
-pywin32==306
-PyYAML==6.0.1
-pyzmq==25.1.1
-referencing==0.30.2
-regex==2023.6.3
+
 requests==2.31.0
-requests-oauthlib==1.3.1
-rich==13.5.2
-rpds-py==0.9.2
-rsa==4.9
-s3transfer==0.6.1
-sacremoses==0.0.53
-safetensors==0.3.1
-scikit-learn==1.3.0
-scipy==1.11.2
-sentencepiece==0.1.99
-six==1.16.0
-sklearn==0.0.post7
-smmap==5.0.0
-sniffio==1.3.1
-stack-data==0.6.2
-starlette==0.37.2
-streamlit==1.25.0
-sympy==1.12
-telebot==0.0.5
-tenacity==8.2.3
-tensorboard==2.13.0
-tensorboard-data-server==0.7.1
-tensorflow==2.13.0
-tensorflow-estimator==2.13.0
-tensorflow-intel==2.13.0
-tensorflow-io-gcs-filesystem==0.31.0
-termcolor==2.3.0
-textdistance==4.5.0
-tflearn==0.5.0
-threadpoolctl==3.2.0
-tokenizers==0.13.3
-toml==0.10.2
-toolz==0.12.0
-torch==2.0.1
-tornado==6.3.3
-tqdm==4.65.0
-traitlets==5.9.0
-transformers==2.1.0
-typing_extensions==4.11.0
-tzdata==2023.3
-tzlocal==4.3.1
-ujson==5.9.0
-uritemplate==4.1.1
-urllib3==1.26.16
+
+scikit-learn==1.4.1.post1
+
+
 uvicorn==0.29.0
-validators==0.21.2
-watchdog==3.0.0
-watchfiles==0.21.0
-wcwidth==0.2.6
-websockets==12.0
-Werkzeug==2.3.7
-wrapt==1.15.0
-zipp==3.16.2
+python-multipart==0.0.9
